refactor: co-locate errors to error generating code
