+++ src/constraint.rs
@@ -1,8 +1,8 @@
-use crate::VersionConstraintError;
 use crate::comparator::Comparator;
 use crate::version::VersionString;
 use std::fmt::{Display, Formatter};
 use std::str::FromStr;
+use thiserror::Error;
 
 /// A single version constraint pairing a [`Comparator`] with a validated [`VersionString`].
 #[derive(Clone, PartialEq, Eq, Debug)]
@@ -49,4 +49,21 @@ impl Display for VersionConstraint {
     fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
         write!(f, "{}{}", self.comparator, self.version)
     }
+}
+
+/// Errors specific to a single constraint within a VLS string.
+#[derive(Error, Debug, PartialEq, Eq)]
+pub enum VersionConstraintError {
+    /// A constraint segment was empty (e.g. from `||`, a leading `|`, or a trailing `|`).
+    #[error("Empty constraint")]
+    EmptyConstraint,
+
+    /// The version part of a constraint was empty (e.g. `>=` without a version).
+    #[error("Empty version in constraint")]
+    EmptyVersion,
+
+    /// The version string contains characters outside the allowed grammar.
+    /// See vls::Vls for more details on the grammar.
+    #[error("Invalid character(s) in version string: {}", .0.iter().map(|c| format!("'{}'", c.escape_default())).collect::<Vec<_>>().join(", "))]
+    InvalidVersionCharacters(Vec<char>),
 }
