+++ src/lib.rs
@@ -1,13 +1,11 @@
 // public api
 pub use comparator::Comparator;
-pub use constraint::VersionConstraint;
-pub use error::{VersionConstraintError, VlsError};
+pub use constraint::{VersionConstraint, VersionConstraintError};
 pub use version::VersionString;
-pub use vls::Vls;
+pub use vls::{Vls, VlsError};
 
 mod comparator;
 mod constraint;
-mod error;
 mod valid_chars;
 mod version;
 mod vls;
